refactor(api-keys): migrate to RSC pattern and CustomAlertModal
Split api-keys-card into server and client components. Migrate all modals to use CustomAlertModal. Move types to feature directory and remove custom hook in favor of server component data fetching.
