+++ ui/components/users/profile/create-api-key-modal.tsx
@@ -2,17 +2,12 @@
 
 import { Button } from "@heroui/button";
 import { Input } from "@heroui/input";
-import {
-  Modal,
-  ModalBody,
-  ModalContent,
-  ModalFooter,
-  ModalHeader,
-} from "@heroui/modal";
+import { ModalFooter } from "@heroui/modal";
 import { useState } from "react";
 
 import { createApiKey } from "@/actions/api-keys/api-keys";
 import { Alert, AlertDescription } from "@/components/ui/alert/Alert";
+import { CustomAlertModal } from "@/components/ui/custom/custom-alert-modal";
 
 import { DEFAULT_EXPIRY_DAYS } from "./api-keys/constants";
 import { calculateExpiryDate } from "./api-keys/utils";
@@ -77,53 +72,64 @@ export const CreateApiKeyModal = ({
   };
 
   return (
-    <Modal isOpen={isOpen} onClose={handleClose} size="lg">
-      <ModalContent>
-        <ModalHeader className="flex flex-col gap-1">
+    <CustomAlertModal
+      isOpen={isOpen}
+      onOpenChange={(open) => !open && handleClose()}
+      title="Create API Key"
+      size="lg"
+    >
+      <div className="flex flex-col gap-4">
+        <Input
+          label="Name"
+          labelPlacement="inside"
+          variant="bordered"
+          placeholder="My API Key"
+          value={name}
+          onChange={(e) => setName(e.target.value)}
+          isRequired
+          description="A descriptive name to identify this API key"
+          classNames={{
+            label: "tracking-tight font-light !text-default-500 text-xs z-0!",
+            input: "text-default-500 text-small",
+          }}
+        />
+
+        <Input
+          label="Expires in (days)"
+          labelPlacement="inside"
+          variant="bordered"
+          type="number"
+          value={expiresInDays}
+          onChange={(e) => setExpiresInDays(e.target.value)}
+          min="1"
+          max="3650"
+          description="Number of days until this key expires (default: 365)"
+          classNames={{
+            label: "tracking-tight font-light !text-default-500 text-xs z-0!",
+            input: "text-default-500 text-small",
+          }}
+        />
+
+        {error && (
+          <Alert variant="destructive">
+            <AlertDescription>{error}</AlertDescription>
+          </Alert>
+        )}
+      </div>
+
+      <ModalFooter>
+        <Button color="default" variant="light" onPress={handleClose}>
+          Cancel
+        </Button>
+        <Button
+          color="success"
+          onPress={handleSubmit}
+          isLoading={isLoading}
+          isDisabled={!name.trim()}
+        >
           Create API Key
-        </ModalHeader>
-        <ModalBody>
-          <div className="flex flex-col gap-4">
-            <Input
-              label="Name"
-              placeholder="My API Key"
-              value={name}
-              onChange={(e) => setName(e.target.value)}
-              isRequired
-              description="A descriptive name to identify this API key"
-            />
-
-            <Input
-              label="Expires in (days)"
-              type="number"
-              value={expiresInDays}
-              onChange={(e) => setExpiresInDays(e.target.value)}
-              min="1"
-              max="3650"
-              description="Number of days until this key expires (default: 365)"
-            />
-
-            {error && (
-              <Alert variant="destructive">
-                <AlertDescription>{error}</AlertDescription>
-              </Alert>
-            )}
-          </div>
-        </ModalBody>
-        <ModalFooter>
-          <Button color="default" variant="light" onPress={handleClose}>
-            Cancel
-          </Button>
-          <Button
-            color="success"
-            onPress={handleSubmit}
-            isLoading={isLoading}
-            isDisabled={!name.trim()}
-          >
-            Create API Key
-          </Button>
-        </ModalFooter>
-      </ModalContent>
-    </Modal>
+        </Button>
+      </ModalFooter>
+    </CustomAlertModal>
   );
 };
