+++ ui/components/users/profile/index.ts
@@ -1,5 +1,6 @@
 export * from "./api-key-success-modal";
 export * from "./api-keys-card";
+export * from "./api-keys-card-client";
 export * from "./create-api-key-modal";
 export * from "./delete-api-key-modal";
 export * from "./edit-api-key-name-modal";
